+++ crates/identity/affinidi-did-common/src/lib.rs
@@ -19,6 +19,8 @@ pub mod builder;
 pub mod did;
 pub mod did_method;
 pub mod document;
+#[cfg(feature = "key-agreement")]
+pub mod key_negotiation;
 pub mod one_or_many;
 pub mod service;
 pub mod verification_method;